+++ scrapydweb/run.py
@@ -2,11 +2,11 @@
 import os
 import sys
 import platform
-from shutil import copyfile, rmtree
-import re
-import argparse
 from subprocess import Popen
+import argparse
+import re
 import json
+from shutil import copyfile, rmtree
 
 from flask import request
 
@@ -16,8 +16,9 @@ from scrapydweb import create_app
 from scrapydweb.__version__ import __version__, __description__, __url__
 
 
+SCRAPYDWEB_SETTINGS_PY = 'scrapydweb_settings_v3.py'
 CWD = os.path.dirname(os.path.abspath(__file__))
-SCRAPYDWEB_SETTINGS_PY = 'scrapydweb_settings_v2.py'
+main_pid = os.getpid()
 
 
 def main():
@@ -25,15 +26,15 @@ def main():
     print(">>> Run 'scrapydweb -h' to get help")
     print(">>> Loading default settings from %s" % os.path.join(CWD, 'default_settings.py'))
     app = create_app()
-
     scrapydweb_settings_py = find_scrapydweb_settings_py()
     if scrapydweb_settings_py:
         print(">>> Overriding custom settings from %s" % scrapydweb_settings_py)
         app.config.from_pyfile(scrapydweb_settings_py)
     else:
+        scrapydweb_settings_py = os.path.join('.', SCRAPYDWEB_SETTINGS_PY)
         try:
-            copyfile(os.path.join(CWD, 'default_settings.py'), os.path.join('.', SCRAPYDWEB_SETTINGS_PY))
-            print(">>> The config file '%s' is copied to your working directory, "
+            copyfile(os.path.join(CWD, 'default_settings.py'), scrapydweb_settings_py)
+            print(">>> The config file '%s' is copied to current working directory, "
                   "and you may custom settings with it" % SCRAPYDWEB_SETTINGS_PY)
         except:
             print(">>> You may copy the file 'default_settings.py' from above path to your working directory, "
@@ -42,16 +43,18 @@ def main():
     args = parse_args(app.config)
     check_args(args)
     update_app_config(app.config, args)
-    # print(app.config)
+    # from pprint import pprint
+    # pprint(app.config)
 
+    print(">>> Main pid: %s" % main_pid)
     if not app.config['DISABLE_CACHE']:
         import atexit
-        proc = start_caching(app.config)
-        print(">>> Caching utf8 and stats files in the background with pid: %s" % proc.pid)
-        atexit.register(kill_child, proc)
+        caching_subprocess = start_caching(app.config)
+        print(">>> Caching utf8 and stats files in the background with pid: %s" % caching_subprocess.pid)
+        atexit.register(kill_child, caching_subprocess)
 
-    print(">>> Visit ScrapydWeb at http://{host}:{port} or http://127.0.0.1:{port}".format(
-        host='IP-OF-THE-HOST-WHERE-SCRAPYDWEB-RUNS-ON', port=app.config['SCRAPYDWEB_PORT']))
+    print(">>> Visit ScrapydWeb at http://{bind}:{port} or http://127.0.0.1:{port}".format(
+        bind='IP-OF-THE-HOST-WHERE-SCRAPYDWEB-RUNS-ON', port=app.config['SCRAPYDWEB_PORT']))
 
 
     username = app.config.get('USERNAME', '')
@@ -64,9 +67,12 @@ def main():
             'SCRAPYD_SERVERS_GROUPS': app.config['SCRAPYD_SERVERS_GROUPS'],
             'SCRAPYD_SERVERS_AUTHS': app.config['SCRAPYD_SERVERS_AUTHS'],
             'DEFAULT_LATEST_VERSION': DEFAULT_LATEST_VERSION,
+            'SCRAPYDWEB_VERSION': __version__,
             'GITHUB_URL': __url__,
             'REQUIRE_LOGIN': True if username and password else False,
-            'HIDE_SCRAPYD_ITEMS': app.config.get('HIDE_SCRAPYD_ITEMS', False),
+            'SHOW_SCRAPYD_ITEMS': app.config.get('SHOW_SCRAPYD_ITEMS', True),
+            'scrapydweb_settings_py': scrapydweb_settings_py,
+            'DAEMONSTATUS_REFRESH_INTERVAL': int(app.config.get('DAEMONSTATUS_REFRESH_INTERVAL', 10)),
         }
 
     # https://stackoverflow.com/questions/34164464/flask-decorate-every-route-at-once
@@ -80,9 +86,9 @@ def main():
                 return authenticate()
 
     # /site-packages/flask/app.py
-    # run(host=None, port=None, debug=None, load_dotenv=True, **options)
-    # Threaded mode is enabled by default.
-    app.run(host=app.config['SCRAPYDWEB_HOST'], port=app.config['SCRAPYDWEB_PORT'])  # , debug=True)
+        # def run(self, host=None, port=None, debug=None, load_dotenv=True, **options):
+        # Threaded mode is enabled by default.
+    app.run(host=app.config['SCRAPYDWEB_BIND'], port=app.config['SCRAPYDWEB_PORT'])  # , debug=True)
 
 
 def find_scrapydweb_settings_py(path='.', prevpath=None):
@@ -98,20 +104,10 @@ def find_scrapydweb_settings_py(path='.', prevpath=None):
 def parse_args(config):
     parser = argparse.ArgumentParser(description='ScrapydWeb -- %s' % __description__)
 
-    default = config.get('SCRAPYD_SERVERS', ['127.0.0.1:6800'])
-    parser.add_argument(
-        '-ss', '--scrapyd_server',
-        # default=default,
-        action='append',
-        help=("default: %s, type '-ss 127.0.0.1 -ss username:password@192.168.123.123:6801#group' "
-              "to set any number of Scrapyd servers to control. "
-              "See 'https://github.com/my8100/scrapydweb/blob/master/scrapydweb/default_settings.py' "
-              "for more help") % (default or ['127.0.0.1:6800'])
-    )
 
-    default = config.get('SCRAPYDWEB_HOST', '0.0.0.0')
+    default = config.get('SCRAPYDWEB_BIND', '0.0.0.0')
     parser.add_argument(
-        '-H', '--host',
+        '--bind',
         default=default,
         help=("default: %s, and 0.0.0.0 makes ScrapydWeb server visible externally, "
               "set to 127.0.0.1 to disable that") % default
@@ -138,12 +134,33 @@ def parse_args(config):
         help="default: %s, the password of basic auth for web UI" % default
     )
 
+
+    default = config.get('SCRAPYD_SERVERS', ['127.0.0.1:6800'])
+    parser.add_argument(
+        '-ss', '--scrapyd_server',
+        # default=default,
+        action='append',
+        help=("default: %s, type '-ss 127.0.0.1 -ss username:password@192.168.123.123:6801#group' "
+              "to set any number of Scrapyd servers to control. "
+              "See 'https://github.com/my8100/scrapydweb/blob/master/scrapydweb/default_settings.py' "
+              "for more help") % (default or ['127.0.0.1:6800'])
+    )
+
+
+    default = config.get('SCRAPY_PROJECTS_DIR', '')
+    parser.add_argument(
+        '--scrapy_projects_dir',
+        default=default,
+        help=("default: %s, set to enable auto eggifying in Deploy page, "
+              "e.g., C:/Users/username/myprojects/ or /home/username/myprojects/") % (default or "''")
+    )
+
     default = config.get('SCRAPYD_LOGS_DIR', '')
     parser.add_argument(
         '--scrapyd_logs_dir',
         default=default,
-        help=("default: %s , set to speed up loading utf8 and stats html, "
-              "e.g. C:/Users/username/logs/ or /home/username/logs/ , "
+        help=("default: %s, set to speed up loading utf8 and stats html, "
+              "e.g., C:/Users/username/logs/ or /home/username/logs/ , "
               "The setting takes effect only when both ScrapydWeb and Scrapyd run on the same machine, "
               "and the Scrapyd server ip is added as '127.0.0.1'. "
               "See 'https://scrapyd.readthedocs.io/en/stable/config.html#logs-dir' "
@@ -183,11 +200,19 @@ def check_args(args):
     password = args.password
     if username or password:
         if not username:
-            sys.exit("!!! username should NOT be empty string: %s" % username)
+            sys.exit("!!! In order to enable basic auth, the username should NOT be empty string: '%s'" % username)
         elif not password:
-            sys.exit("!!! password should NOT be empty string: %s" % password)
+            sys.exit("!!! In order to enable basic auth, the password should NOT be empty string: '%s'" % password)
         else:
             print(">>> Enabling basic auth username/password: '%s'/'%s'" % (username, password))
+
+
+    scrapy_projects_dir = args.scrapy_projects_dir
+    if scrapy_projects_dir:
+        if not os.path.isdir(scrapy_projects_dir):
+            sys.exit("!!! scrapy_projects_dir NOT found: %s" % scrapy_projects_dir)
+        else:
+            print(">>> Using scrapy_projects_dir: %s" % scrapy_projects_dir)
 
     scrapyd_logs_dir = args.scrapyd_logs_dir
     if scrapyd_logs_dir:
@@ -211,10 +236,11 @@ def update_app_config(config, args):
 
     config.update(dict(
         SCRAPYD_SERVERS=args.scrapyd_server,
-        SCRAPYDWEB_HOST=args.host,
+        SCRAPYDWEB_BIND=args.bind,
         SCRAPYDWEB_PORT=args.port,
         USERNAME=args.username,
         PASSWORD=args.password,
+        SCRAPY_PROJECTS_DIR=args.scrapy_projects_dir,
         SCRAPYD_LOGS_DIR=args.scrapyd_logs_dir,
     ))
 
@@ -265,7 +291,8 @@ def start_caching(config):
         sys.executable,
         os.path.join(CWD, 'cache.py'),
 
-        '127.0.0.1' if config['SCRAPYDWEB_HOST'] == '0.0.0.0' else config['SCRAPYDWEB_HOST'],
+        str(main_pid),
+        '127.0.0.1' if config['SCRAPYDWEB_BIND'] == '0.0.0.0' else config['SCRAPYDWEB_BIND'],
         str(config['SCRAPYDWEB_PORT']),
         config['USERNAME'],
         config['PASSWORD'],
@@ -275,14 +302,23 @@ def start_caching(config):
         str(config['CACHE_REQUEST_INTERVAL']),
     ]
 
-    kwargs = {}
-    if platform.system() == 'Windows':
+    # 'Windows':
+        # AttributeError: module 'signal' has no attribute 'SIGKILL'
         # ValueError: preexec_fn is not supported on Windows platforms
-        pass
+    # macOS('Darwin'):
+        # subprocess.SubprocessError: Exception occurred in preexec_fn.
+        # OSError: dlopen(libc.so.6, 6): image not found
+    if platform.system() == 'Linux':
+        kwargs = dict(preexec_fn=on_parent_exit('SIGKILL')) # 'SIGTERM' 'SIGKILL'
+        try:
+            caching_subprocess = Popen(args, **kwargs)
+        except Exception as err:
+            print(err)
+            caching_subprocess = Popen(args)
     else:
-        kwargs.update(preexec_fn=on_parent_exit('SIGKILL')) # 'SIGTERM'
-    proc = Popen(args, **kwargs)
-    return proc
+        caching_subprocess = Popen(args)
+
+    return caching_subprocess
 
 
 if __name__ == '__main__':