+++ tests/conftest.py
@@ -1,15 +1,21 @@
 # coding: utf8
+import os
+
 import pytest
 
 from scrapydweb import create_app
 from scrapydweb.vars import DEFAULT_LATEST_VERSION
 
 
+CWD = os.path.dirname(os.path.abspath(__file__))
+
+
 @pytest.fixture
 def app():
     app = create_app({
         'TESTING': True,
-        'SCRAPYD_SERVERS': ['127.0.0.1:6800']
+        'SCRAPYD_SERVERS': ['127.0.0.1:6800'],
+        'SCRAPY_PROJECTS_DIR': os.path.join(CWD, 'data'),
     })
 
     @app.context_processor
@@ -19,6 +25,7 @@ def app():
             'SCRAPYD_SERVERS_GROUPS': ['fakegroup' for s in app.config['SCRAPYD_SERVERS']],
             'SCRAPYD_SERVERS_AUTHS': [('fakeusername', 'fakepassword') for s in app.config['SCRAPYD_SERVERS']],
             'DEFAULT_LATEST_VERSION': DEFAULT_LATEST_VERSION,
+            'DAEMONSTATUS_REFRESH_INTERVAL': 0,
         }
 
     yield app